backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
easyocr
transformers
torch
torchvision
opencv-python-headless
numpy
spacy
thefuzz